refactor: put splash screen as root index
